backend/requirements.txt
Flask==2.2.5
requests==2.31.0
celery==5.3.1
redis==4.5.5
gunicorn==20.1.0
mysql-connector-python==8.0.33
python-dotenv
pytz==2024.1
pandas
openpyxl
